WIP updated server.py with app config
