+++ pyproject.toml
@@ -1,14 +1,15 @@
 [project]
 name = "comfydock"
-version = "0.2.1"
+version = "0.2.2"
 description = "ComfyDock CLI"
 readme = "README.md"
 requires-python = ">=3.12"
 dependencies = [
     "click>=8.1.8",
-    "comfydock-server>=0.2.1",
+    "comfydock-server>=0.3.0",
     "packaging>=24.2",
     "python-dotenv>=1.0.1",
+    "setuptools>=80.9.0",
 ]
 
 [project.scripts]
