+++ setup.py
@@ -1,0 +1,51 @@
+from setuptools import setup, find_packages
+
+with open("README.md", "r", encoding="utf-8") as fh:
+    long_description = fh.read()
+
+setup(
+    name="comfydock",
+    version="0.1.0",
+    author="Your Name",
+    author_email="your.email@example.com",
+    description="ComfyDock CLI - Manage ComfyUI Docker environments",
+    long_description=long_description,
+    long_description_content_type="text/markdown",
+    url="https://github.com/yourusername/comfydock-cli",
+    packages=find_packages(),
+    package_data={
+        "comfydock": ["config.json"],
+    },
+    include_package_data=True,
+    classifiers=[
+        "Development Status :: 3 - Alpha",
+        "Intended Audience :: Developers",
+        "License :: OSI Approved :: MIT License",
+        "Operating System :: OS Independent",
+        "Programming Language :: Python :: 3",
+        "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: 3.9",
+        "Programming Language :: Python :: 3.10",
+        "Programming Language :: Python :: 3.11",
+    ],
+    python_requires=">=3.8",
+    install_requires=[
+        "click>=8.0.0",
+        "comfydock-server>=0.1.4",
+        "requests>=2.25.0",
+        "packaging>=20.0",
+        "python-dotenv>=0.19.0",
+    ],
+    extras_require={
+        "dev": [
+            "pytest>=6.0",
+            "black>=21.0",
+            "flake8>=3.8",
+        ],
+    },
+    entry_points={
+        "console_scripts": [
+            "comfydock=comfydock.cli:main",
+        ],
+    },
+)
